+++ src/symmetric/prf.rs
@@ -9,7 +9,7 @@ pub trait Pseudorandom {
     fn gen<R: Rng>(rng: &mut R) -> Self::Key;
 
     /// Apply the one-way function to an epoch and an index
-    fn apply(key: &Self::Key, epoch: u64, index: u64) -> Self::Output;
+    fn apply(key: &Self::Key, epoch: u32, index: u64) -> Self::Output;
 }
 
 pub mod hashprf;
